Typing consistency applied for 64-bit plus IBB file reader

+++ LUMA/inc/GridObj.h
@@ -58,9 +58,9 @@ private :
 
 	// Start and end indices of corresponding coarse level
 	// When using MPI these values are local to a particular coarse grid
-	size_t CoarseLimsX[2];
-	size_t CoarseLimsY[2];
-	size_t CoarseLimsZ[2];
+	int CoarseLimsX[2];
+	int CoarseLimsY[2];
+	int CoarseLimsZ[2];
 
 	// 1D arrays
 public :
@@ -136,7 +136,7 @@ public :
 	// LBM operations
 	void LBM_multi(bool IBM_flag);		// Launch the multi-grid kernel
 	void LBM_collide();					// Apply collision + 1 overload for equilibrium calculation
-	double LBM_collide(size_t i, size_t j, size_t k, size_t v, size_t M_lim, size_t K_lim);
+	double LBM_collide(int i, int j, int k, int v, int M_lim, int K_lim);
 	void LBM_kbcCollide(int i, int j, int k, int M_lim, int K_lim, IVector<double>& f_new);		// KBC collision operator
 	void LBM_stream();							// Stream populations
 	void LBM_macro();							// Compute macroscopic quantities + 1 overload for single site
